# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NCH Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [estimate, deposit_received, in_progress, invoiced, paid, complete, closed]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job status or details
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/receipts:
    get:
      operationId: listJobReceipts
      tags: [receipts]
      summary: List receipts for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Log a receipt for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptBody"
      responses:
        "201":
          description: Created receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /jobs/{jobId}/generate-estimate:
    post:
      operationId: generateEstimate
      tags: [invoices]
      summary: Generate estimate PDF for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDocument"
  /jobs/{jobId}/generate-invoice:
    post:
      operationId: generateInvoice
      tags: [invoices]
      summary: Generate invoice PDF for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDocument"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Job:
      type: object
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        client:
          type: string
        address:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [estimate, deposit_received, in_progress, invoiced, paid, complete, closed]
        estimateAmount:
          type: number
        depositAmount:
          type: number
        totalCosts:
          type: number
        marginPercent:
          type: number
        isOverBudget:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - jobNumber
        - client
        - address
        - status
        - estimateAmount
        - depositAmount
        - totalCosts
        - isOverBudget
        - createdAt
        - updatedAt
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            receipts:
              type: array
              items:
                $ref: "#/components/schemas/Receipt"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/JobNote"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusChange"
          required:
            - receipts
            - notes
            - statusHistory
    CreateJobBody:
      type: object
      properties:
        client:
          type: string
          enum: [BSMK, Coastal Management LLC]
        address:
          type: string
        description:
          type: string
        estimateAmount:
          type: number
      required:
        - client
        - address
        - description
        - estimateAmount
    UpdateJobBody:
      type: object
      properties:
        status:
          type: string
          enum: [estimate, deposit_received, in_progress, invoiced, paid, complete, closed]
        description:
          type: string
        estimateAmount:
          type: number
        note:
          type: string
    Receipt:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        amount:
          type: number
        category:
          type: string
          enum: [materials, labor, subcontractor, equipment_tools, vehicle_fuel, other]
        vendorName:
          type: string
        notes:
          type: string
        photoUrl:
          type: string
        driveFileId:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
      required:
        - id
        - jobId
        - amount
        - category
        - createdAt
        - createdBy
    CreateReceiptBody:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
          enum: [materials, labor, subcontractor, equipment_tools, vehicle_fuel, other]
        vendorName:
          type: string
        notes:
          type: string
        photoBase64:
          type: string
      required:
        - amount
        - category
    JobNote:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        note:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - note
        - author
        - createdAt
    StatusChange:
      type: object
      properties:
        id:
          type: integer
        fromStatus:
          type: string
        toStatus:
          type: string
        changedBy:
          type: string
        note:
          type: string
        changedAt:
          type: string
          format: date-time
      required:
        - id
        - toStatus
        - changedBy
        - changedAt
...
    GeneratedDocument:
      type: object
      properties:
        invoiceId:
          type: integer
        invoiceNumber:
          type: string
        pdfUrl:
          type: string
        driveFileId:
          type: string
      required:
        - invoiceId
        - invoiceNumber
...